+++ master/buildbot/test/regressions/test_oldpaths.py
@@ -193,7 +193,3 @@ class OldImportPaths(unittest.TestCase):
         from buildbot.steps.source import Monotone
         assert Monotone
 
-    def test_steps_source_BK(self):
-        from buildbot.steps.source import BK
-        assert BK
-